# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface FragranceNotes {
  top: string[];
  heart: string[];
  base: string[];
}

export type FragranceConcentration =
  (typeof FragranceConcentration)[keyof typeof FragranceConcentration];

export const FragranceConcentration = {
  EDP: "EDP",
  EDT: "EDT",
  Parfum: "Parfum",
  EDC: "EDC",
} as const;

export interface Fragrance {
  id: string;
  name: string;
  house: string;
  year: number;
  concentration: FragranceConcentration;
  accords: string[];
  notes: FragranceNotes;
  /**
   * @minimum 1
   * @maximum 5
   */
  longevity: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  sillage: number;
  price_gbp: number;
}

export interface FindDupesRequest {
  fragranceName: string;
  priceCeiling?: number | null;
}

export interface DupeResult {
  name: string;
  house: string;
  similarity_pct: number;
  price_gbp: number;
  price_delta: number;
  accords: string[];
}

export interface ContextBody {
  fragranceName: string;
  weatherTemp: number;
  weatherDesc: string;
  occasion: string;
  timeOfDay: string;
  ownedFragrances: string[];
}

export interface ContextPick {
  rank: number;
  name: string;
  house: string;
  reason: string;
  match_pct: number;
}

export interface ScoreBody {
  fragranceName: string;
  ownedFragrances?: string[];
  budget?: number | null;
}

export interface ScoreBreakdown {
  accord_compatibility: number;
  community_longevity: number;
  batch_consistency: number;
  price_value: number;
}

export type RiskFlagLevel = (typeof RiskFlagLevel)[keyof typeof RiskFlagLevel];

export const RiskFlagLevel = {
  warn: "warn",
  info: "info",
  ok: "ok",
} as const;

export interface RiskFlag {
  level: RiskFlagLevel;
  message: string;
}

export type BlindBuyScoreVerdict =
  (typeof BlindBuyScoreVerdict)[keyof typeof BlindBuyScoreVerdict];

export const BlindBuyScoreVerdict = {
  Strong_buy: "Strong buy",
  Buy: "Buy",
  Try_first: "Try first",
  Avoid: "Avoid",
} as const;

export interface BlindBuyScore {
  overall_score: number;
  breakdown: ScoreBreakdown;
  verdict: BlindBuyScoreVerdict;
  risk_flags: RiskFlag[];
  recommendation: string;
}

export interface WeatherData {
  temp_c: number;
  description: string;
  humidity: number;
}

export type UserProfileBudget =
  | (typeof UserProfileBudget)[keyof typeof UserProfileBudget]
  | null;

export const UserProfileBudget = {
  under_50: "under_50",
  "50_150": "50_150",
  "150_300": "150_300",
  no_limit: "no_limit",
} as const;

export interface UserProfile {
  ownedFragrances: string[];
  budget: UserProfileBudget;
}

export interface SaveProfileRequest {
  ownedFragrances: string[];
  budget?: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type SearchFragrancesParams = {
  /**
   * Search query (name or house)
   */
  q: string;
};

export type GetWeatherParams = {
  lat: string;
  lon: string;
};