# lib/api-zod/src/generated/types/contextBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ContextBody {
  fragranceName: string;
  weatherTemp: number;
  weatherDesc: string;
  occasion: string;
  timeOfDay: string;
  ownedFragrances: string[];
}