# lib/api-zod/src/generated/types/contextPick.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ContextPick {
  rank: number;
  name: string;
  house: string;
  reason: string;
  match_pct: number;
}