# lib/api-zod/src/generated/types/fragrance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { FragranceConcentration } from "./fragranceConcentration";
import type { FragranceNotes } from "./fragranceNotes";

export interface Fragrance {
  id: string;
  name: string;
  house: string;
  year: number;
  concentration: FragranceConcentration;
  accords: string[];
  notes: FragranceNotes;
  /**
   * @minimum 1
   * @maximum 5
   */
  longevity: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  sillage: number;
  price_gbp: number;
}