# lib/api-zod/src/generated/types/riskFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFlagLevel } from "./riskFlagLevel";

export interface RiskFlag {
  level: RiskFlagLevel;
  message: string;
}