# lib/api-zod/src/generated/types/saveProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveProfileRequest {
  ownedFragrances: string[];
  budget?: string | null;
}