# lib/api-zod/src/generated/types/scoreBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBreakdown {
  accord_compatibility: number;
  community_longevity: number;
  batch_consistency: number;
  price_value: number;
}