# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileBudget } from "./userProfileBudget";

export interface UserProfile {
  ownedFragrances: string[];
  budget: UserProfileBudget;
}