# lib/api-zod/src/generated/types/weatherData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scentinel Fragrance Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherData {
  temp_c: number;
  description: string;
  humidity: number;
}